% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appLog}
\alias{appLog}
\title{App log test dataset.}
\format{
A data frame with 20 rows and 8 variables:
\describe{
  \item{id}{character id}
  \item{time}{character timestamp}
  \item{user}{user name}
  \item{name}{user full name}
  \item{group}{users group/registry}
  \item{role}{users role}
  \item{resh_id}{users organization}
  \item{message}{log message}
}
}
\usage{
appLog
}
\description{
A dataset containing test entries for the application log.
}
\keyword{datasets}
