% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{autoReport}
\alias{autoReport}
\alias{autoReportUI}
\alias{autoReportInput}
\alias{autoReportServer}
\alias{autoReportApp}
\title{Shiny modules and helper functions for registry auto reports}
\usage{
autoReportUI(id)

autoReportInput(id)

autoReportServer(id, registryName, type, reports = NULL, orgs = NULL)

autoReportApp(
  registryName = "rapbase",
  type = "subscription",
  reports = NULL,
  orgs = NULL
)
}
\arguments{
\item{id}{Character string shiny module id}

\item{registryName}{Character string registry name key. Must correspond to
the R package name}

\item{type}{Character string defining the type of auto reports. Must be one
of \code{c("subscription", "dispatchment", "bulletin")}}

\item{reports}{List of reports that will be provided as automated reports.
Describe further...}

\item{orgs}{List of named organizations and values. Describe further...}
}
\value{
Shiny objects, mostly. Helper functions may return other stuff too.
}
\description{
These shiny modules may be used to set up auto reporting from registries at
Rapporteket
}
