% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportFormatInput}
\alias{autoReportFormatInput}
\title{Report Format Selection UI}
\usage{
autoReportFormatInput(id)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}
}
\value{
A Shiny UI object.
}
\description{
User interface module for selecting report output format.
}
