% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportFormatServer}
\alias{autoReportFormatServer}
\title{Report Format Selection Server}
\usage{
autoReportFormatServer(id)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}
}
\value{
Reactive value containing the selected report format.
}
\description{
Server module for selecting report output format.
}
