% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportInput}
\alias{autoReportInput}
\title{Auto Report Configuration UI}
\usage{
autoReportInput(id)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}
}
\value{
A Shiny UI object.
}
\description{
User interface module for configuring and creating auto reports.
}
