% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportOrgInput}
\alias{autoReportOrgInput}
\title{Organization Selection UI}
\usage{
autoReportOrgInput(id)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}
}
\value{
A Shiny UI object.
}
\description{
User interface module for selecting an organization used as data source
for dispatchments and bulletins.
}
