% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportOrgServer}
\alias{autoReportOrgServer}
\title{Organization Selection Server}
\usage{
autoReportOrgServer(id, orgs)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}

\item{orgs}{Named list containing organization names and ids.}
}
\value{
A list containing:
\describe{
\item{name}{Reactive organization name.}
\item{value}{Reactive organization id.}
}
}
\description{
Server module for organization selection.
}
