% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportServer2}
\alias{autoReportServer2}
\title{Alias for autoReportServer}
\usage{
autoReportServer2(...)
}
\arguments{
\item{...}{Arguments passed directly to [autoReportServer()].}
}
\value{
Same result as [autoReportServer()].
}
\description{
Backward compatible wrapper around [autoReportServer()].
}
