% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportUI}
\alias{autoReportUI}
\title{Auto Report Main UI}
\usage{
autoReportUI(id)
}
\arguments{
\item{id}{Character string providing the Shiny module id.}
}
\value{
A Shiny UI object.
}
\description{
User interface module for displaying existing auto reports and related
information.
}
