% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTools.R
\name{createDb}
\alias{createDb}
\title{Create a Database}
\usage{
createDb(dbName, dbType = "mysql", force = FALSE)
}
\arguments{
\item{dbName}{A character string specifying the name of the database to be
created.}

\item{dbType}{A character string specifying the type of the database.
  Default is `"mysql"`. Other supported types may depend on the
implementation.}

\item{force}{A logical value indicating whether to overwrite an existing
database with the same name. Default is `FALSE`.}
}
\value{
The function does not return a value but creates a database as a
side effect.
}
\description{
This function creates a new database with the specified name and type.
}
\examples{
\dontrun{
createDb("myDatabase") # Creates a MySQL database named "myDatabase"
createDb("myDatabase", dbType = "sqlite") # Creates an SQLite database
createDb("myDatabase", force = TRUE) # Forces creation, overwriting if exists
}

}
