% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{createLogDb}
\alias{createLogDb}
\title{Create a logging database}
\usage{
createLogDb(dbKey)
}
\arguments{
\item{dbKey}{Character string with the key to a corresponding entry for the
database in dbConfig.yml. Please also make sure that the r.raplog.key
property in rapbaseConfig.yml is given the exact same value.}
}
\value{
Invisibly TRUE
}
\description{
Internal function that crates a database to be used for logging. Will return
an error if the database already exists.
}
\keyword{internal}
