% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{createLogDbTabs}
\alias{createLogDbTabs}
\title{Create tables for log entries in a database}
\usage{
createLogDbTabs()
}
\value{
Invisibly TRUE
}
\description{
Internal function that crates a database tables to be used for logging. Will
return an error if the table(s) already exists.
}
\keyword{internal}
