% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{deleteAutoReport}
\alias{deleteAutoReport}
\title{Delete existing report from config/db}
\usage{
deleteAutoReport(autoReportId, target = "db")
}
\arguments{
\item{autoReportId}{String providing the auto report unique id}

\item{target}{List of autoreports in file or database}
}
\description{
Delete existing report from config/db
}
\seealso{
\code{\link{createAutoReport}}
}
