% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{.getFun}
\alias{.getFun}
\title{Provide explicit reference to function for do.call}
\usage{
.getFun(x)
}
\arguments{
\item{x}{string with explicit reference, i.e. 'package::function'}
}
\value{
value of the exported 'function' in 'package'
}
\description{
Provide explicit reference to function for do.call
}
