% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{.testAutoReport}
\alias{.testAutoReport}
\title{Simple test of automated report}
\usage{
.testAutoReport(aNum = 1, aChar = "a", anExp = Sys.Date(), bulletin = 0)
}
\arguments{
\item{aNum}{a number}

\item{aChar}{a character}

\item{anExp}{an expression}

\item{bulletin}{Integer defining if report is of type bulletin (1) or not
(0). Set to 0 by default}
}
\value{
A simple message listing the contents of the arguments
}
\description{
Simple test of automated reporting from definitions provided in a yaml
config file
}
\examples{
.testAutoReport()
}
