% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExport.R
\name{export}
\alias{export}
\alias{exportUCInput}
\alias{exportUCServer}
\alias{exportUCApp}
\alias{selectListPubkey}
\alias{exportDb}
\alias{queryToRdsFile}
\alias{downloadBusyJs}
\title{Shiny modules providing GUI and server logic for Export}
\usage{
exportUCInput(id)

exportUCServer(
  id,
  dbName,
  teamName = NULL,
  eligible = shiny::reactiveVal(TRUE)
)

exportUCApp(dbName = "rapbase")

selectListPubkey(pubkey)

exportDb(dbName, compress = FALSE, session)

queryToRdsFile(dbName, query, compress = FALSE, session)

downloadBusyJs(btnId, idle = "Last ned!", busy = "Laster ned...")
}
\arguments{
\item{id}{Character string module ID}

\item{dbName}{Character string database name}

\item{teamName}{Character string defining the github team name containing
members allowed to export the database. Default value is \code{dbName}.}

\item{eligible}{Logical defining if the module should be allowed to work at
full capacity. This might be useful when access to module products should be
restricted. Default is TRUE, \emph{i.e.} no restrictions.}

\item{pubkey}{Character vector with public keys}

\item{compress}{Logical if the output RDS file should be
compressed using gzip}

\item{session}{Shiny session object}

\item{query}{Character string SQL query to execute}

\item{btnId}{Button id (already namespaced if in module)}

\item{idle}{Label when idle}

\item{busy}{Label when busy}
}
\value{
Shiny objects, mostly. Helper functions may return other stuff too.
}
\description{
Functions for registries that wants to implement exporting of registry
databases, \emph{e.g.} for local development purposes. Also includes
relevant helper functions
}
\examples{
## client user interface function
ui <- shiny::fluidPage(
  shiny::sidebarLayout(
    shiny::sidebarPanel(
      exportUCInput("test"),
    ),
    shiny::mainPanel(
      NULL
    )
  )
)

## server function
server <- function(input, output, session) {
  exportUCServer("test", dbName = "rapbase")
}

## run the shiny app in an interactive environment
if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
