% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportApp.R
\name{exportApp}
\alias{exportApp}
\title{Shiny app with database export functionality}
\usage{
exportApp(repoName = "", registryName = "data")
}
\arguments{
\item{repoName}{Character string registry name key, corresponding to
github team name}

\item{registryName}{Character string registry name key, can be used to
specify name of database if needed.}
}
\description{
Shiny app with database export functionality
}
