% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleExport.R
\name{exportGuide}
\alias{exportGuide}
\alias{exportGuideUI}
\alias{exportGuideServer}
\alias{exportGuideApp}
\alias{exportGuideServer2}
\title{Shiny modules providing the Export Guide}
\usage{
exportGuideUI(id)

exportGuideServer(id, dbName)

exportGuideServer2(id, dbName)

exportGuideApp()
}
\arguments{
\item{id}{Character string module ID}

\item{dbName}{Character string registry name key}
}
\value{
Functions ui and server representing the (module) app
}
\description{
Shiny modules providing the Export Guide
}
\examples{
ui <- shiny::fluidPage(
  exportGuideUI("exportGuide")
)

server <- function(input, output, session) {
  exportGuideServer("exportGuide", "test")
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
