% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFigFilType.R
\name{figtype}
\alias{figtype}
\title{Funksjon som angir hvilket format en figur skal lagres i.}
\usage{
figtype(outfile = "", width = 3 * 595, height = 3 * 595, res = 3 *
  72, pointsizePDF = 11, fargepalett = "BlaaOff")
}
\arguments{
\item{outfile}{Navn på fil som skal skrives til. Tom tekststreng skriver til skjerm}

\item{width}{Bredde på output ved gitt oppløsning}

\item{height}{Høyde på output ved gitt oppløsning}

\item{res}{Oppløsningen på raster-output}

\item{pointsizePDF}{Skrifstørrelse i pdf-output}

\item{fargepalett}{Hvilken fargepalett skal brukes}
}
\value{
UtFigFil List med en del figurparametre
}
\description{
Formatet er spesifisert gjennom endinga til filnavnet
Tillatt: png, pdf, jpg, bmp, tif, svg og ps
Velge farger: BlaaRapp, BlaaHNpms287, GronnHNpms624, GronnHNpms342
}
\keyword{norgast}
\keyword{rapporteket}
