% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getConfig}
\alias{getConfig}
\title{Get configuration for package, if any}
\usage{
getConfig(fileName = "dbConfig.yml", packageName = "rapbase")
}
\arguments{
\item{fileName}{String providing configuration file base name}

\item{packageName}{String providing the package name}
}
\value{
A list of (yaml) configuration
}
\description{
Try to obtain yaml-formatted configuration placed either as given by the
environment variable R_RAP_CONFIG_PATH or as provided by the package itself.
If none can be found the function exits with an error
}
\examples{
getConfig()
}
