% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{getContainerPrivileges}
\alias{getContainerPrivileges}
\title{IN CONSTRUCTION: Get environmental variables for for container apps.}
\usage{
getContainerPrivileges(group, unit = NULL)
}
\arguments{
\item{group}{Character string providing the name of the app R package name.
The term "group" is used to relate to the environmental variable
SHINYPROXY_USERGROUPS that corresponds to the apps a given user can access.}

\item{unit}{Integer providing the look-up unit id. Default value is NULL in
which case all privileges for \code{group} are returned.}
}
\value{
List of privileges
}
\description{
For apps running as containers particular environment variables must be
defined for an orderly handling of dynamic user privileges. This function
makes use of environmental variables defined by shinyproxy to provide
available privileges for the shiny application
}
