% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getRapPackages}
\alias{getRapPackages}
\title{Get all installed Rapporteket packages}
\usage{
getRapPackages()
}
\value{
Character vector of packages names
}
\description{
Get all installed packages that depends on 'rapbase' which itself will not
be reported
}
\examples{
## Relevant only in a Rapporteket-like context
if (isRapContext()) {
  getRapPackages()
}
}
