% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{getRegs}
\alias{getRegs}
\title{Provide vector of registries (\emph{i.e.} their R packages) in config}
\usage{
getRegs(config)
}
\arguments{
\item{config}{list of configuration for automated reports}
}
\value{
character vector of registry (package) names
}
\description{
Provide vector of registries (\emph{i.e.} their R packages) in config
}
