% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{getSessionData}
\alias{getSessionData}
\title{Get session data}
\usage{
getSessionData()
}
\value{
A list of relevant log fields
}
\description{
Internal function providing session data relevant to logging.
}
\keyword{internal}
