% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetShinyUserGroups.R
\name{getShinyUserGroups}
\alias{getShinyUserGroups}
\title{Get user groups from a shiny session object}
\usage{
getShinyUserGroups(shinySession, testCase = FALSE)
}
\arguments{
\item{shinySession}{List shiny session object}

\item{testCase}{Logical defining to use shinySessionInfo on a test case.
For test cases user privileges are provided by a query defined in the
url of the shiny application request. Default FALSE}
}
\value{
String user groups
}
\description{
Get user groups from a shiny session object
}
\examples{
\donttest{
# Requires a valid shiny session object
getShinyUserGroups(shinySessionObject)
}

}
\seealso{
\code{\link{getShinyUserName}}, \code{\link{getShinyUserReshId}},
\code{\link{getShinyUserRole}}
}
