% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetShinyUserName.R
\name{getShinyUserName}
\alias{getShinyUserName}
\title{Get user name from a shiny session object}
\usage{
getShinyUserName(shinySession, testCase = FALSE)
}
\arguments{
\item{shinySession}{List shiny session object}

\item{testCase}{Logical defining to use shinySessionInfo on a test case.
For test cases user privileges are provided by a query defined in the
url of the shiny application request. Default FALSE}
}
\value{
String user name
}
\description{
Get user name from a shiny session object
}
\examples{
\donttest{
# Requires a valid shiny session object
getShinyUserName(shinySessionObject)
}

}
\seealso{
\code{\link{getShinyUserName}}, \code{\link{getShinyUserGroups}},
\code{\link{getShinyUserReshId}}, \code{\link{getShinyUserRole}}
}
