% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetShinyUserReshId.R
\name{getShinyUserReshId}
\alias{getShinyUserReshId}
\title{Get user resh_id from a shiny session object}
\usage{
getShinyUserReshId(shinySession, testCase = FALSE)
}
\arguments{
\item{shinySession}{List shiny session object}

\item{testCase}{Logical defining to use shinySessionInfo on a test case.
For test cases user privileges are provided by a query defined in the
url of the shiny application request. Default FALSE}
}
\value{
String user resh_id
}
\description{
Get user resh_id from a shiny session object
}
\examples{
\donttest{
# Requires a valid shiny session object
getShinyUserReshId(shinySessionObject)
}

}
\seealso{
\code{\link{getShinyUserName}}, \code{\link{getShinyUserGroups}},
\code{\link{getShinyUserRole}}
}
