% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helligdager2008til2022.R
\docType{data}
\name{Helligdager2008til2022}
\alias{Helligdager2008til2022}
\title{Offisielle norske helligdager 2008 til 2022}
\format{En data frame med 195 rader og 2 kolonner
\describe{
  \item{Dato}{Dato helligdagen faller på, i datoformat}
  \item{Helligdag}{Hvilken helligdag det gjelder}
}}
\source{
\url{http://www.norskkalender.no/}
}
\usage{
Helligdager2008til2022
}
\description{
Et datasett med norske røddager 2008 til 2022. Søndager 
er ikke inkludert i listen.
}
\keyword{datasets}
