% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleNavbarWidget.R
\name{howWeDealWithPersonalData}
\alias{howWeDealWithPersonalData}
\title{Render text in pop-up}
\usage{
howWeDealWithPersonalData(..., callerPkg = NULL)
}
\arguments{
\item{...}{Further arguments, currently not used}
}
\value{
fragment html info text
}
\description{
Render text on how Rapporteket deals with personal data
}
