% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/InstallGithubPackage.R
\name{installGithubPackage}
\alias{installGithubPackage}
\title{Install packages from GitHub at Rapporteket}
\usage{
installGithubPackage(packageName, branchName)
}
\arguments{
\item{packageName}{String Name of the package}

\item{branchName}{String Name of the branch to use}
}
\value{
message String containing logged entries from the function
}
\description{
Install and report results
}

