% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isPkgRapReg}
\alias{isPkgRapReg}
\title{Test if a package is part of Rapporteket}
\usage{
isPkgRapReg(pkg)
}
\arguments{
\item{pkg}{String providing the package name}
}
\value{
Logical TRUE if pkg depends on 'rapbase', FALSE if not
}
\description{
Test if an installed package is linked to Rapporteket based on
dependency to the package 'rapbase'
}
\examples{
# returns FALSE, rapbase has no explicit dependency to itself
isPkgRapReg("rapbase")

}
\seealso{
\code{\link{getRapPackages}} on how to list all packages that
depend om 'rapbase'
}
