% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRegData.R
\name{loadRegData}
\alias{loadRegData}
\alias{describeRegistryDb}
\alias{nlinesRegistryDb}
\title{Provider of data for registries at Rapporteket}
\usage{
loadRegData(registryName, query, dbType = "mysql")

describeRegistryDb(registryName, tabs = c())

nlinesRegistryDb(registryName, tabs = c())
}
\arguments{
\item{registryName}{String providing the name of the database to connect to.
If it is "data" it will use the MYSQL_DB_DATA environment variable, if it is
"autoreport" it will use the MYSQL_DB_AUTOREPORT environment variable, and
if it is "raplog" it will use the MYSQL_DB_LOG environment variable. If
none of these are set, it will use the name provided.}

\item{query}{String SQL query to obtain the data}

\item{dbType}{String Type of db to query, currently "mysql" (default) and
"mssql"}

\item{tabs}{Character vector for optional definition of tables to describe.
Defaults to an empty vector in which case all tables are used}
}
\value{
data frame containing registry data or a list with table names and
corresponding fields with attributes
}
\description{
Generic to registries, provide reporting data obtained from sql databases
Underlying this function is rapbase::RapporteketDbConnection
}
