% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appLog.R
\name{logShinyInputChanges}
\alias{logShinyInputChanges}
\title{Wrapper around logger::log_shiny_input_changes}
\usage{
logShinyInputChanges(...)
}
\arguments{
\item{...}{Arguments passed to logger::log_shiny_input_changes function}
}
\description{
Wrapper around logger::log_shiny_input_changes
}
