% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{makeAutoReportTab}
\alias{makeAutoReportTab}
\title{Make table of automated reports}
\usage{
makeAutoReportTab(session, type = "subscription", mapOrgId = NULL)
}
\arguments{
\item{session}{A shiny session object}

\item{type}{Character string defining the type of auto reports to tabulate.
Must be one of 'subscription', 'dispatchment' or 'bullitin'. Default value
set to 'subscription'}

\item{mapOrgId}{Data frame containing the two columns 'name' and 'id'
corresponding to unique name and id of organizations. Defult is NULL in
which case the ids provided in auto report data will be used. In case
mapOrgId is not NULL but no id match is found the id fond in the auto
report data will also be used}
}
\value{
Matrix providing a table to be rendered in a shiny app
}
\description{
Make a table to be rendered in a shiny app providing automated reports
from a given user or registry as obtained from the shiny session
object provided.
}
\details{
Each table record (line) represents a uniqely defined automated report.
For each line two shiny action buttons are provided to allow
for editing and deleting uniqe dispatchments, respectively. For applications
implementing this table observing events on these action buttons may be used
to allow users to manage automated reports by GUI. The
action buttons for editing and deleting are provided with the static input
ids \emph{edit_button} and \emph{del_button} and upon clicking the
\emph{button} part of their ids will change to the uniqe id of the
dispatchment. Hence, a GUI call for editing a dispatchment can be catched by
\code{shiny::observeEvent("edit_button")} and within this event the
dispatchment id is obtained by collecting the string after the underscore,
\emph{e.g.} \code{strsplit(input$edit_button, "_")[[1]][2]}.

Take a look at the example shiny server function in
\href{https://github.com/Rapporteket/rapRegTemplate}{rapRegTemplate} on how
this function may be implemented
}
