% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMessage.R
\name{MakeMessage}
\alias{MakeMessage}
\title{Provide simple formatting of a sequential message}
\usage{
MakeMessage(story, message, newLine = TRUE, newPara = FALSE,
  timeStamp = TRUE)
}
\arguments{
\item{story}{String Text that the message is to be added to}

\item{message}{String Message to add}

\item{newLine}{Logical [TRUE] Start message at a new line. TRUE by default}

\item{newPara}{Logical [FALSE] Start message at new paragraph. Default FALSE}

\item{timeStamp}{Logical [TRUE] Add a time stamp to the message. TRUE by
default}
}
\value{
story String The whole story...
}
\description{
Wrap strings into a message for simple presentation
}
