% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{makeRegDispatchmentTab}
\alias{makeRegDispatchmentTab}
\title{Make table of registry dispatchments of reports}
\usage{
makeRegDispatchmentTab(session, mapOrgId = NULL)
}
\arguments{
\item{session}{A shiny session object}

\item{mapOrgId}{Data frame containing the two columns 'name' and 'id'
corresponding to unique name and id of organizations. Defult is NULL in
which case the ids provided in auto report data will be used. In case
mapOrgId is not NULL but no id match is found the id fond in the auto
report data will also be used}
}
\value{
Matrix providing a table to be rendered in a shiny app
}
\description{
Make a table to be rendered in a shiny app providing the active
dispatchments from a given registry as obtained from the shiny session
object provided
}
