% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStandardTable.R
\name{makeStandardTable}
\alias{makeStandardTable}
\alias{mst}
\title{Make standard table for rmarkdown reports}
\usage{
mst(
  tab,
  col_names = colnames(tab),
  type = "latex",
  cap = "",
  label = "",
  digs = 0,
  align = NULL,
  fs = 8,
  lsd = FALSE
)
}
\arguments{
\item{tab}{Data frame or matrix represetnting the table}

\item{col_names}{Character vector with column names. Defaults
\code{colnames(tab)}}

\item{type}{Character string defining output, either 'html' or 'latex'.
Default is 'latex'}

\item{cap}{Character string with table caption. Empty string by default}

\item{label}{Character string defining the label in case the table needs to
be referenced elsewhere in the overall document. For instance, setting this
to 'my_table' the corresponding inline rmarkdown reference to use is
\code{\\@ref(tab:my_table)}. Please note that for this to work for both
LaTex and HTML the bookdown document processing functions must be used,
\emph{i.e.} bookdown:pdf_document2() and bookdown::html_document2(),
respectively. Default value is \code{knitr::opts_current$get("label")}.}

\item{digs}{Numeric number of digits to use. = by default}

\item{align}{Character vector specifying column alignment in the LaTeX way,
\emph{e.g.} \code{c("l", "c", "r")} will align the first column to the left,
center the second and right-aling the last one. Default is NULL in which case}

\item{fs}{Numeric providing the font size. Only apply for LaTeX output.
Default value is 8}

\item{lsd}{Logical if table is to be scaled down. Only apply for LaTeX
output. FALSE by default}
}
\value{
Character string containing RMarkdown table code
}
\description{
Functions that will return tables used in reports.
}
\details{
\code{mst()} creates RMarkdown code for creating standard tables.

\code{prettyTab()} creates a table with cells containing percentages of
totals and observations optinally with a marginal row count
}
\examples{
mst(tab = mtcars[1:10, ])
}
