% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{orgList2df}
\alias{orgList2df}
\title{Convert Organization List to Data Frame}
\usage{
orgList2df(orgs)
}
\arguments{
\item{orgs}{Named list of organizations.}
}
\value{
Data frame with columns `name` and `id`.
}
\description{
Internal utility for converting a named organization list into
a data frame.
}
\keyword{internal}
