% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/PushGist.R
\name{PushGist}
\alias{PushGist}
\title{Push some markdown to gist}
\usage{
PushGist(mdFile, gistId = "", PAT = "")
}
\arguments{
\item{mdFile}{String absolute file name and path of source Rmd-file. For
instance provided as a \code{system.file} expression.}

\item{gistId}{String unike id to the gist to be updated. Defaults to an
empty string which means that a gist will be created. If provided the
corresponding gist will be updated}

\item{PAT}{String Personal Access Token in case the gist is not public.
Defaults to empty}
}
\description{
Based on Rmd documents, push the resulting md to defined gist
}

