% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/rapbase.R
\docType{package}
\name{rapbase}
\alias{rapbase}
\alias{rapbase-package}
\title{rapbase: A package for generic R-resources at Rapporteket.}
\description{
Provide common functions and resources for registry speciffic
 R-packages at Rapportekt. Can potentially contain alot of stuff, also
 not directly related to R...
}
\section{rapbase functions}{

RunNoweb
LibFigFilType
RapporteketDbConnection
}

