% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RapporteketDbConnection.R
\name{rapCloseDbConnection}
\alias{rapCloseDbConnection}
\title{Close down data connection handle}
\usage{
rapCloseDbConnection(con, drv)
}
\arguments{
\item{con}{Open connection object that is to be closed}

\item{drv}{DBIDriver object that is to be removed}
}
\description{
Close down data connection handle
}

