% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageLog.R
\name{raplogManager}
\alias{raplogManager}
\title{Management of shiny application and report logging at Rapporteket}
\usage{
raplogManager(overSize = 1024 * 1000)
}
\arguments{
\item{overSize}{Integer size in bytes from where larger files will be listed
as candidates for archiving. Not used by this function \emph{per se} but
rather passed on to \code{logsOverSize()}. Normally, its value does not need
to be set but may be convenient in rare cases (such as full scale tests).
Default value set to 1 Mb (1024 * 1000)}
}
\value{
Silently exits after successful file operations
}
\description{
A tool to aid archiving and cleanup of logs containing data on use of shiny
application and reports at Rapporteket. This function is ment to be called
from a scheduler (\emph{e.g.} chron) directly or as part of a process chain
initiated by a scheduler
}
\examples{
\donttest{
# Depend on the R_RAP_CONFIG_PATH being defined and that relevant
# configuration is present
raplogManager()
}

}
\seealso{
To become orchestrator
\code{\link[rapbase:runAutoReport]{runAutoReport}} in package \emph{rapbase}
}
