% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{rapOpenDbConnection}
\alias{rapOpenDbConnection}
\title{Provide connection handle for data source at Rapporteket}
\usage{
rapOpenDbConnection(dbName, dbType = "mysql")
}
\arguments{
\item{dbName}{String providing the name of the database to connect to. If it
is "data" it will use the MYSQL_DB_DATA environment variable, if it is
"autoreport" it will use the MYSQL_DB_AUTOREPORT environment variable, and
if it is "raplog" it will use the MYSQL_DB_LOG environment variable. If
none of these are set, it will use the name provided.}

\item{dbType}{String providing type of data source, one of
  "mysql" and "sqlite". Defaults to "mysql". "mssql" is not supported
anymore.}
}
\value{
A named list of con and drv representing the db connection handle and
  driver, respectively.
}
\description{
Generic to registries, handle the data source connections, including
usernames and passwords needed to open these connections
}
