% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{rapTheme}
\alias{rapTheme}
\title{Theme of the app.}
\usage{
rapTheme(theme = "flatly", version = 3)
}
\arguments{
\item{theme}{Name of the theme. See bslib::bootswatch_themes()
for available themes.}

\item{version}{Version of bootstrap to use (3, 4 or 5).}
}
\value{
a Bootstrap theme object
}
\description{
This is a wrapper around bslib::bs_theme
}
