% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{readAutoReportData}
\alias{readAutoReportData}
\title{Read automated report metadata}
\usage{
readAutoReportData()
}
\value{
a table of autoreport data
}
\description{
Read automated report metadata
}
\examples{
readAutoReportData()
}
