% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{readLog}
\alias{readLog}
\title{Read log entries}
\usage{
readLog(type, name = "", app_id = NULL)
}
\arguments{
\item{type}{Character string defining which log to request data from. Must be
one of \code{c("app", "report")}.}

\item{name}{Character string with registry filter. Default value is an empty
string. WARNING: this is not used in the function, but is kept for
compatibility with the old function signature. It is recommended to use
\code{app_id} instead, which is the new parameter for filtering by
application ID.}

\item{app_id}{An identifier for a particular registry. Default value is NULL,
in which case no action is taken. If value is provided, the log is filtered
to show only entries matching chosen app_id.}
}
\value{
A data frame of log entries
}
\description{
Internal function that provide log entries
}
\keyword{internal}
