% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegDelay.R
\name{RegDelay}
\alias{RegDelay}
\title{Provide registration delay for OpenQReg quality registries}
\usage{
RegDelay(years, registryName, registrationFormName, peculiarity = FALSE)
}
\arguments{
\item{years}{integer vector with years for results and grouping}

\item{registryName}{String Key/name of registry to get data from}

\item{registrationFormName}{String Name of the registry form where dates
are to be collected from}
}
\value{
data frame with registry name, quartiles 1, 2 and 3 and number of
observations for each year
}
\description{
Provide registration delay in median number of days (in addition to first
and third quartile, and N) grouped by years
}
\seealso{
Data to this function is provided by \code{\link{RegDelayData}}
and \code{\link{RegDelayDataPeculiar}}.
}

