% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegDelayData.R
\name{RegDelayData}
\alias{RegDelayData}
\title{Provide registration delay data from OpenQReg quality registries}
\format{Return a data frame with two variables:
\describe{
\item{year}{the year for the event (from db field \emph{HovedDato})}
\item{daysDiff}{the difference in days between event (from db field
\emph{HovedDato}) and the last time the registration form was saved
(from db field \emph{SistLagretDato})}
}}
\usage{
RegDelayData(registryName, registrationFormName)
}
\arguments{
\item{registryName}{String Key/name of registry to get data from}

\item{registrationFormName}{String Name of the registry form where dates
are to be collected from}
}
\value{
regDelayData data frame
}
\description{
Provides delay between event (main date) and registration as a data frame
}
\details{
For the query these conditions apply:
\describe{
\item{SkjemaStatus = 1}{use Only finished registrations}
\item{SkjemaNavn = registrationFormName}{use only registry form name given
by the input argument \emph{registrationFormName}. For surgical registries
this is usually the form "Operasjon".}
}
}
\seealso{
This function is used by \code{\link{RegDelay}}.
}

