% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegDelayDataPeculiar.R
\name{RegDelayDataPeculiar}
\alias{RegDelayDataPeculiar}
\title{Provide registration delay data from quality registries other than OpenQReg}
\format{Return a data frame with two variables:
\describe{
\item{year}{the year for the event}
\item{daysDiff}{the difference in days between event and the last time the
registration form was saved (preferably)}
}}
\usage{
RegDelayDataPeculiar(registryName)
}
\arguments{
\item{registryName}{String Key/name of registry to get data from}
}
\value{
regDelayData data frame
}
\description{
Provides delay between event (main date) and registration as a data frame.
This function handles each of the defined registries separately.
}
\seealso{
This function takes care of those registries than can not be
handled by \code{\link{RegDelayData}}.
}

