% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegDelayReport.R
\name{RegDelayReport}
\alias{RegDelayReport}
\title{Make report data on registration delays across OpenQReg registries}
\usage{
RegDelayReport(years, registryKeyNames, registrationFormNames, peculiarity)
}
\arguments{
\item{years}{Integer Vector of years to group delays}

\item{registryKeyNames}{String Vector of registry names corresponding to the
key used for within registry packages}

\item{registrationFormNames}{String Vector of registration form names where
dates are to be collected from. Elements must correspond to
\emph{registryKeyNames}}

\item{peculiarity}{Logical vector, TRUE for none-OpenQReg registries.
Elements must correspond to \emph{registryKeyNames}}
}
\value{
regDelays Data frame of median delays
}
\description{
Make report data on registration delays across OpenQReg registries
}
\seealso{
This function use \code{\link{RegDelay}}
}

