% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{runBulletin}
\alias{runBulletin}
\title{Run bulletin auto reports}
\usage{
runBulletin()
}
\value{
Whatever \code{runAutoReport()} might provide
}
\description{
This is a wrapper for \code{runAutoReport()} to issue bulletins.
}
