% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{runNoweb}
\alias{runNoweb}
\title{runNoweb}
\usage{
runNoweb(nowebFileName, packageName, weaveMethod = "knitr")
}
\arguments{
\item{nowebFileName}{Basename of the noweb file, \emph{e.g.} 'myFile.Rnw'.}

\item{packageName}{Name of the package containing noweb file(s)}

\item{weaveMethod}{Method to apply for weaving. Currently available are
'Sweave' and 'knitr', default to the latter.}
}
\description{
Function to run noweb file contained in a package. Assume all noweb files
of the package are placed flat under the \emph{inst} directory
}
