% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{runShinyApp starts a shiny application in a package}
\usage{
runShinyApp(appName, appsDirectoryName, packageName)
}
\arguments{
\item{appName}{String providing the name of the shiny application}

\item{appsDirectoryName}{String providing the within-package-path to its
shiny application(s). Assumingly, in the source tree for a given R package
an app will be found under \emph{inst/shinyApps/anApp}. In this case,
\emph{appsDirectoryName} must be set to \emph{'shinyApps'}.}

\item{packageName}{String providing the name of the package in which the
shiny app resides}
}
\description{
runShinyApp start a shiny application residing within an R package. First
attempt is to keep thin function generic (across packages). It might not
prove wise and future implementation might be per-package.
}
\details{
Idea and code (partly) taken from Dean Attali's R Blog
(http://www.r-bloggers.com/supplementing-your-r-package-with-a-shiny-app/)
Gratitude.
}

