% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{sanitizeLog}
\alias{sanitizeLog}
\title{Sanitize log entries that have reached end of life}
\usage{
sanitizeLog(eolDays = 730)
}
\arguments{
\item{eolDays}{Number of days to keep log entries. Entries older than this
will be removed. Default value is 730 days (2 years).}
}
\value{
NULL on success
}
\description{
Function that removes log entries older than a given number of days.
}
