% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{selectByOrganization}
\alias{selectByOrganization}
\title{Select data on one organization from config (list)}
\usage{
selectByOrganization(config, organization)
}
\arguments{
\item{config}{list of configuration for automated reports}

\item{organization}{string giving the exact organization}
}
\value{
list with config for reports belonging to organization
}
\description{
Pick all config corresponding to a given organization (of the report)
}
