% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{selectByOwner}
\alias{selectByOwner}
\title{Select data on one owner from config (list)}
\usage{
selectByOwner(config, owner)
}
\arguments{
\item{config}{list of configuration for automated reports}

\item{owner}{string giving the exact name owner}
}
\value{
list with config for reports belonging to owner
}
\description{
Pick all config corresponding to a given owner (of the report)
}
