% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{selectByReg}
\alias{selectByReg}
\title{Select data on one registry from config (list)}
\usage{
selectByReg(config, reg)
}
\arguments{
\item{config}{list of configuration for automated reports}

\item{reg}{string giving the exact name of the R package for the registry}
}
\value{
list with config for registry reg
}
\description{
Pick all config corresponding to a given registry. Registry name is not
given as such, but rather as its corresponding R package name. Hence, a
registry must be given as the name of its R package
}
