% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{selectByType}
\alias{selectByType}
\title{Select data by type from config (list)}
\usage{
selectByType(config, type)
}
\arguments{
\item{config}{List of configuration for automated reports}

\item{type}{Character string giving the type}
}
\value{
list with config for reports of a given type
}
\description{
Pick reports of a given type
}
