% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{sendEmail}
\alias{sendEmail}
\title{Send email from Rapporteket}
\usage{
sendEmail(conf, to, subject, text, attFile = NULL)
}
\arguments{
\item{conf}{List containing (Rapporteket) config such as sender email
address, SMTP server url and port number}

\item{to}{Character vector containing email addresses. May also contain
full names like '\code{Jane Doe <janed@nowhere.com>}'}

\item{subject}{Character string providing email subject. The string is
converted within this function to conform to RFC 1342}

\item{text}{Character string providing the plain email text}

\item{attFile}{Character string providing the full file path to an
attachment. Default is NULL in which case no attachment is made}
}
\value{
Invisible sending of email
}
\description{
This function can be used to send email from within R at Rapporteket. It
relies on (and must hence be provided) specific settings through local
configuration to work properly.
}
