% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{setContainerEnv}
\alias{setContainerEnv}
\title{Set environmental variables for for container apps.}
\usage{
setContainerEnv(shinySession, unit = NULL)
}
\arguments{
\item{shinySession}{A shiny session object}

\item{unit}{Integer providing the look-up unit id. Default value is NULL in
which case the first viable option will be used}
}
\value{
Invisible NULL on success
}
\description{
For apps running as containers particular environment variables must be
defined for an orderly handling of dynamic user privileges. This function
makes use of information stored in the shiny session environment and in
environmental variables defined by shinyproxy to set (new) environment
variables APP_ORG and APP_ROLE that defines the current role and organization
for the app.
}
