% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySessionInfo.R
\name{shinySessionInfo}
\alias{shinySessionInfo}
\title{Provide elements from a shiny session object}
\usage{
shinySessionInfo(shinySession, entity, testCase = FALSE)
}
\arguments{
\item{shinySession}{List shiny session object}

\item{entity}{String defining the element to return. One of 'user',
'groups', 'resh_id' or 'role'}

\item{testCase}{Logical defining to use shinySessionInfo on a test case.
For test cases user privileges are provided by a query defined in the
url of the shiny application request. Default FALSE}
}
\value{
String of single user data element
}
\description{
Extracts elements from a shiny session object relevant for user data. This
function will normally be used via its helper functions.
}
\seealso{
\code{\link{getShinyUserName}}, \code{\link{getShinyUserGroups}},
\code{\link{getShinyUserReshId}}, \code{\link{getShinyUserRole}}
}
