% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagingData.R
\name{stagingData}
\alias{stagingData}
\alias{pathStagingData}
\alias{saveStagingData}
\alias{loadStagingData}
\title{Staging data functions}
\usage{
pathStagingData(registryName, dir)

saveStagingData(
  registryName,
  dataName,
  data,
  dir = Sys.getenv("R_RAP_CONFIG_PATH")
)

loadStagingData(registryName, dataName, dir = Sys.getenv("R_RAP_CONFIG_PATH"))
}
\arguments{
\item{registryName}{Character string providing the registry name}

\item{dir}{Character string providing the path to where the staging data
directory resides. Default value is \code{Sys.getenv("R_RAP_CONFIG_PATH")}.}

\item{dataName}{Character string providing the data set name.}

\item{data}{A data object such as a data.frame to be stored as
\code{dataName}.}
}
\value{
\code{pathStagingData()} returns a character string with the path to
the staging directory of \code{registryName}. If its parent directory
(\code{dir}) does not exists an error is returned. \code{saveStagingData()}
returns the data object (\code{data}), invisibly. \code{loadStagingData()}
returns the data object corresponding to the name given upon saving
(\code{dataName}). If the requested data set does not exist FALSE is
returned.
}
\description{
These functions are used to handle staging data at Rapporteket.
}
\examples{
## Check and make the path to staging data
path <- pathStagingData(registryName = "myReg", dir = tempdir())

## Prep test data
registryName <- "myReg"
dataName <- "testData"
data <- mtcars
dir <- tempdir()

## Save data for staging
sageStagingData(registryName, dataName, data, dir)

## Retrieve data set from staging
loadStagingData(registryName, dataName, dir)
}
