% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleStats.R
\name{statsGuide}
\alias{statsGuide}
\alias{statsGuideUI}
\alias{statsGuideServer}
\alias{statsGuideApp}
\title{Shiny modules providing the Stats Guide}
\usage{
statsGuideUI(id)

statsGuideServer(id, registryName)

statsGuideApp()
}
\arguments{
\item{id}{Character string module ID}

\item{registryName}{Character string registry name key}
}
\value{
Functions ui and server representing the (module) app
}
\description{
Shiny modules providing the Stats Guide
}
\examples{
ui <- shiny::fluidPage(
  statsGuideUI("statsGuide")
)

server <- function(input, output, session) {
  statsGuideServer("statsGuide", "test")
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
