% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test dataset.}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{id}{int id}
  \item{someText}{char text}
  \item{someInt}{int integer}
  \item{someBigInt}{int big integer}
  \item{someFloat}{num float number}
  \item{someTime}{time date}
}
}
\usage{
testdata
}
\description{
A dataset containing test entries.
}
\keyword{datasets}
