% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{theme}
\alias{theme}
\title{DO NOT USE THIS FUNCTION, use rapTheme instead.
This is only for backward compatibility.}
\usage{
theme(...)
}
\arguments{
\item{...}{rapTheme arguments}
}
\value{
a Bootstrap theme object
}
\description{
DO NOT USE THIS FUNCTION, use rapTheme instead.
This is only for backward compatibility.
}
