% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{title}
\alias{title}
\title{DO NOT USE THIS FUNCTION, use regTitle instead.
This is only for backward compatibility.}
\usage{
title(...)
}
\arguments{
\item{...}{regTitle arguments}
}
\value{
a div containing the logo and the title
}
\description{
DO NOT USE THIS FUNCTION, use regTitle instead.
This is only for backward compatibility.
}
