% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{unitAttribute}
\alias{unitAttribute}
\title{Get unit attributes from an access tree file}
\usage{
unitAttribute(unit, what, file = NULL, path = Sys.getenv("R_RAP_CONFIG_PATH"))
}
\arguments{
\item{unit}{Integer providing the look-up unit id}

\item{what}{Character string defining what to return for the given unit id}

\item{file}{Character string file name of the JSON file. Default values is
NULL in which case the corresponding value from rapbaseConfig.yml will be
used.}

\item{path}{Character string file path of the JSON file. Default value is
\code{Sys.getenv("R_RAP_CONFIG_PATH")}.}
}
\value{
The corresponding value of 'what'.
}
\description{
Obtain organization unit attributes from an access tree JSON file
}
