% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{upgradeAutoReportData}
\alias{upgradeAutoReportData}
\title{Upgrade auto reports}
\usage{
upgradeAutoReportData(config)
}
\arguments{
\item{config}{List of auto report configuration}
}
\value{
List of (upgraded) auto report configuration
}
\description{
Upgrade auto report config as new features emerge. Currently, the type
definition is added and set to 'subscription' that historically has been
the only type used
}
