% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{userAttribute}
\alias{userAttribute}
\alias{getUserEmail}
\alias{getUserFullName}
\alias{getUserGroups}
\alias{getUserName}
\alias{getUserPhone}
\alias{getUserReshId}
\alias{getUserRole}
\title{Get user attributes}
\usage{
getUserEmail(shinySession = NULL, group = NULL)

getUserFullName(shinySession = NULL, group = NULL)

getUserGroups(shinySession = NULL, group = NULL)

getUserName(shinySession = NULL, group = NULL)

getUserPhone(shinySession = NULL, group = NULL)

getUserReshId(shinySession = NULL, group = NULL)

getUserRole(shinySession = NULL, group = NULL)
}
\arguments{
\item{shinySession}{A shiny session object. Default value is NULL}

\item{group}{Character string providing the name of the app R package name.
The term "group" is used to relate to the environmental variable
SHINYPROXY_USERGROUPS that corresponds to the apps a given user can access.
Default value is NULL but should always be set when shiny app is run as a
shinyproxy container.}
}
\value{
String with user attribute
}
\description{
These are helper function for \code{\link{userInfo}}. When used without a
shiny session object calls to these functions is made without any arguments.
If redefining contexts is needed, please use \code{\link{userInfo}} instead.
}
\examples{
\donttest{
# Requires a valid shiny session object
try(getUserEmail())
try(getUserEmail(shinySessionObject))
}
}
