% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{userInfo}
\alias{userInfo}
\title{Provide user attributes based on environment context}
\usage{
userInfo(entity)
}
\arguments{
\item{entity}{String defining the element to return. Currently, one of
'user', groups', 'resh_id', 'role', 'email', 'full_name' or 'phone'.}
}
\value{
String of single user data element
}
\description{
Extracts elements from either config, url (shiny), shiny session or
environmental variables relevant for user data such as name, group, role and
org id (\emph{e.g.} resh id). Source of info is based on environment context
and can be controlled by altering the default settings for which contexts
that will apply for the various sources of user data. This function will
normally be used via its helper functions (see below).
}
\seealso{
\code{\link{getUserName}}, \code{\link{getUserGroups}},
  \code{\link{getUserReshId}}, \code{\link{getUserRole}}
}
