% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{appendLog}
\alias{appendLog}
\title{Append a log record}
\usage{
appendLog(event, name, target, format)
}
\arguments{
\item{event}{data.frame of one record holding the fields of whatever that
is to be logged}

\item{name}{String defining the name of the log}

\item{target}{String defining where the log should go. Currently, 'file' is
the only option}

\item{format}{String defining the whatever format available given 'taget'.
Currently, the only option here is 'csv'}
}
\value{
Provides a new record in the log. If the log does not exist a new
one is created before appending the new record
}
\description{
Append a record to the log. Normally, this function will not be called from
outside this package
}
