% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAutoReport.R
\name{autoReportServer}
\alias{autoReportServer}
\title{Auto Report Server Module}
\usage{
autoReportServer(
  id,
  registryName,
  type,
  org = NULL,
  paramNames = shiny::reactiveVal(c("")),
  paramValues = shiny::reactiveVal(c("")),
  reports = NULL,
  orgs = NULL,
  eligible = shiny::reactiveVal(TRUE),
  freq = "month",
  user,
  runAutoReportButton = FALSE
)
}
\arguments{
\item{id}{Character string providing the shiny module id.}

\item{registryName}{Character string with registry package name.}

\item{type}{One of `"subscription"`, `"dispatchment"` or `"bulletin"`.}

\item{org}{Reactive organization id.}

\item{paramNames}{Reactive vector of parameter names.}

\item{paramValues}{Reactive vector of parameter values.}

\item{reports}{Report metadata list.}

\item{orgs}{Named list of organizations.}

\item{eligible}{Reactive logical indicating module availability.}

\item{freq}{Default report frequency.}

\item{user}{User metadata reactives.}

\item{runAutoReportButton}{Logical indicating if testing button should be
shown.}
}
\value{
A Shiny server module.
}
\description{
Creates and manages automated report subscriptions, dispatchments,
and bulletins.
}
\details{
The `reports` argument must be a list where each entry represents
one report configuration:

\describe{
  \item{synopsis}{Description of the report.}
  \item{fun}{Exported report function name.}
  \item{paramNames}{Function argument names.}
  \item{paramValues}{Corresponding argument values.}
}
}
