% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{deleteAutoReport}
\alias{deleteAutoReport}
\title{Delete existing report from config}
\usage{
deleteAutoReport(autoReportId)
}
\arguments{
\item{autoReportId}{String providing the auto report unique id}
}
\description{
Delete existing report from config
}
\seealso{
\code{\link{createAutoReport}}
}
