% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{.testAutoReport}
\alias{.testAutoReport}
\title{Simple test of automated report}
\usage{
.testAutoReport(aNum = 1, aChar = "a", anExp = Sys.Date())
}
\arguments{
\item{aNum}{a number}

\item{aChar}{a character}

\item{anExp}{an expression}
}
\value{
A simple message listing the contents of the arguments
}
\description{
Simple test of automated reporting from definitions provided in a yaml
config file
}
\examples{
.testAutoReport()
}
