% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{exportUCInput}
\alias{exportUCServer}
\alias{exportUCApp}
\alias{selectListPubkey}
\alias{exportDb}
\title{Shiny modules providing GUI and server logic for Export}
\usage{
exportUCInput(id)

exportUCServer(id, registryName)

exportUCApp(registryName = "rapbase")

selectListPubkey(pubkey)

exportDb(registryName, compress = FALSE, session)
}
\arguments{
\item{id}{Character string module ID}

\item{registryName}{Character string registry name key}

\item{pubkey}{Character vector with public keys}

\item{compress}{Logical if export data is to be compressed (using gzip).
FALSE by default.}

\item{session}{Shiny session object}
}
\value{
Shiny objects, mostly. Helper functions may return other stuff too.
}
\description{
Functions for registries that wants to implement exporting of registry
databases, \emph{e.g.} for local development purposes. Also includes
relevant helper functions
}
