% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{selectListPubkey}
\alias{exportDb}
\title{Tools for exporting data}
\usage{
selectListPubkey(pubkey)

exportDb(registryName, compress = FALSE, session)
}
\arguments{
\item{pubkey}{Character vector with public keys}

\item{registryName}{Character string registry name key}

\item{compress}{Logical if export data is to be compressed (using gzip).
FALSE by default.}

\item{session}{Shiny session object}
}
\value{
Some data structure with info obtained from the github api
}
\description{
Functions tools for registries that want implement exporting of registry
databases for local development purposes.
}
