% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportApp.R
\name{exportApp}
\alias{exportApp}
\title{Shiny app with database export functionality}
\usage{
exportApp(teamName = "", dbName = "data")
}
\arguments{
\item{teamName}{Character string, corresponding to
github team name}

\item{dbName}{Character string, can be used to
specify name of database if needed. Defaults to "data",
which will work for most registries.}
}
\description{
Shiny app with database export functionality
}
