% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportApp.R
\name{exportApp}
\alias{exportApp}
\title{Shiny app with database export functionality}
\usage{
exportApp(registryName = "")
}
\arguments{
\item{registryName}{Character string registry name key, corresponding to
github team name}
}
\description{
Shiny app with database export functionality
}
