% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportGuide.R, R/statsGuide.R
\name{exportGuide}
\alias{exportGuide}
\alias{exportGuideUI}
\alias{exportGuideServer}
\alias{exportGuideApp}
\alias{statsGuideServer}
\alias{statsGuideApp}
\title{Shiny modules providing the Export Guide}
\usage{
exportGuideUI(id)

exportGuideServer(id, registryName)

exportGuideApp()

statsGuideServer(id, registryName)

statsGuideApp()
}
\arguments{
\item{id}{Character string module ID}

\item{registryName}{Character string registry name key}
}
\value{
Functions ui and server representing the (module) app
}
\description{
Shiny modules providing the Export Guide
}
\examples{
ui <- shiny::fluidPage(
  exportGuideUI("exportGuide")
)

server <- function(input, output, session) {
  exportGuideServer("exportGuide", "test")
}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
