% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/LibFigFilType.R
\name{figtype}
\alias{figtype}
\title{Funksjon som angir hvilket format en figur skal lagres i.}
\usage{
figtype(outfile = "", res = 72, width = 555, height = 555,
  pointsize = 12, fargepalett = "BlaaOff")
}
\arguments{
\item{outfile}{??}

\item{res}{??}

\item{width}{??}

\item{height}{??}

\item{pointsize}{??}

\item{fargepalett}{??}
}
\description{
Formatet er spesifisert gjennom endinga til filnavnet
Tillatt: png, pdf, jpg, bmp, tif, svg og ps
Velge farger: BlaaRapp, BlaaHNpms287, GronnHNpms624, GronnHNpms342
}
\keyword{norgast}
\keyword{rapporteket}

