% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireInTheHole.R
\name{fireInTheHole}
\alias{fireInTheHole}
\title{Kick off functions at Rapporteket}
\usage{
fireInTheHole(flipPeriod = FALSE)
}
\arguments{
\item{flipPeriod}{Logical only used for testing. FALSE by default}
}
\description{
This function will normally be executed by a cron daemon. Once started this
function will nest through schedule functions defined in a configuration
file, \emph{e.g.} "rapbaseConfig.yml".
}
\details{
This is a crontab example running fireInTheHole() every night at 01 hours,
Monday through Friday and with emails suppressed:
\preformatted{0  1 * * 1-5 Rscript -e 'rapbase::fireInTheHole()' >/dev/null
2>&1}
}
\examples{
\donttest{
# Depends on the env var R_RAP_CONFIG_PATH being properly set
fireInTheHole()
}

}
