% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{getDbConfig}
\alias{getDbConfig}
\title{Get database connection configuration}
\usage{
getDbConfig(registryName = "MYSQL_DB_DATA")
}
\arguments{
\item{registryName}{String id used for the registry in global configuration
file from which information on the database connection is provided}
}
\value{
A list with name, user, password and host of the db connection.
}
\description{
Get database connection configuration
}
\keyword{internal}
