% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{getSessionData}
\alias{getSessionData}
\title{Get session data}
\usage{
getSessionData(session, group = NULL)
}
\arguments{
\item{session}{A shiny session object}

\item{group}{Character string providing the name of the app R package name.
The term "group" is used to relate to the environmental variable
SHINYPROXY_USERGROUPS that corresponds to the apps a given user can access.}
}
\value{
A list of relevant log fields
}
\description{
Internal function providing session data relevant to logging.
}
\keyword{internal}
