% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howWeDealWithPersonalData.R
\name{howWeDealWithPersonalData}
\alias{howWeDealWithPersonalData}
\title{Render text in pop-up}
\usage{
howWeDealWithPersonalData(session)
}
\arguments{
\item{session}{A shiny session object used to personalize the text}
}
\value{
html pop-up text
}
\description{
Render text on how Rapporteket deals witn personal data
}
