% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isRapContext}
\alias{isRapContext}
\title{Rapporteket context}
\usage{
isRapContext()
}
\value{
Logical if system has a defined Rapporteket context
}
\description{
Call to this function will return TRUE when run on a system where the
environment variable \code{R_RAP_INSTANCE} is set to either "DEV", "TEST",
"QA" or "PRODUCTION" and FALSE otherwise
}
\examples{
isRapContext()

}
