% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRegData.R
\name{LoadRegData}
\alias{LoadRegData}
\title{Provider of data for registries at Rapporteket}
\usage{
LoadRegData(registryName, query, dbType = "mysql")
}
\arguments{
\item{registryName}{String Name of the registry as defined in dbConfig.yml}

\item{query}{String SQL query to obtain the data}

\item{dbType}{String Type of db to query, currently "mysql" (default) and "mssql"}
}
\value{
RegData dataframe Registry data
}
\description{
Generic to registries, provide reporting data obtained from sql databases
Underlying this function is rapbase::RapporteketDbConnection
}
