% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appLog.R
\name{loggerSetup}
\alias{loggerSetup}
\title{Settings for logging as json}
\usage{
loggerSetup(
  usernameEnv = "SHINYPROXY_USERNAME",
  appidEnv = "SHINYPROXY_APPID",
  hooks = TRUE
)
}
\arguments{
\item{usernameEnv}{Global variable containing user name}

\item{appidEnv}{Global variable containing application name}

\item{hooks}{Logical defining if hooks for automatic logging should be set}
}
\description{
Every info, warning and error will be logged in json format.
}
