% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appLog.R
\name{logShinyInputChanges}
\alias{logShinyInputChanges}
\title{Wrapper around logger::log_shiny_input_changes}
\usage{
logShinyInputChanges(input)
}
\arguments{
\item{input}{passed from Shiny's server}
}
\description{
Wrapper around logger::log_shiny_input_changes
}
