% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{makeLogRecord}
\alias{makeLogRecord}
\title{Make a log record}
\usage{
makeLogRecord(content, format = "csv")
}
\arguments{
\item{content}{A named list of values to be logged}

\item{format}{String defining the format of the log record. Supported
values: 'csv' (default)}
}
\value{
A formatted log entry
}
\description{
Add default values and make a formatted log record
}
\examples{
makeLogRecord(list(msg="This is a test"))
}
