% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{makeUserSubscriptionTab}
\alias{makeUserSubscriptionTab}
\title{Make table of subscriptions of reports}
\usage{
makeUserSubscriptionTab(session)
}
\arguments{
\item{session}{A shiny session object}
}
\value{
Matrix providing a table to be rendered in a shiny app
}
\description{
Make a table to be rendered in a shiny app providing the active
subscriptions of a given user within a given registry which are
both collected from the shiny session object provided
}
