% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{memoryTitle}
\alias{memoryTitle}
\title{Title on the top left of the app, including the logo and memory usage}
\usage{
memoryTitle(title)
}
\arguments{
\item{title}{Title of the app}
}
\value{
a tagList containing the logo, the title, and memory usage
}
\description{
Title on the top left of the app, including the logo and memory usage
}
\examples{
app_ui <- function() {
  shiny::navbarPage(
    title = memoryTitle("My App")
  )
}
}
