% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{noOptOutOk}
\alias{noOptOutOk}
\title{Provide a no-opt-out ok message}
\usage{
noOptOutOk()
}
\value{
String with possible state of mind (in Norwegian) once left with no
options
}
\description{
To be applied for user input when there is actually no choice :-)
}
\examples{
noOptOutOk()
}
