% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PushGist.R
\name{PushGist}
\alias{PushGist}
\title{Push some markdown to gist}
\usage{
PushGist(mdFile, githubUserName = "")
}
\arguments{
\item{mdFile}{String absolute file name and path of source Rmd-file. For
instance provided as a \code{system.file} expression.}

\item{githubUserName}{String github user name for whom the gist will be
published}
}
\description{
Based on Rmd documents, push the resulting md to defined gist
}
