% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapbase.R
\docType{package}
\name{rapbase}
\alias{rapbase}
\title{rapbase: Base Functions and Resources for Rapporteket}
\description{
Provide common functions and resources for registry specific
R-packages at Rapporteket. This packages is relevant for developers of
packages/registries at Rapporteket
}
