% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapporteketDbConnection.R
\name{rapCloseDbConnection}
\alias{rapCloseDbConnection}
\title{Close down data connection handle}
\usage{
rapCloseDbConnection(con)
}
\arguments{
\item{con}{Open connection object that is to be closed}
}
\description{
Close down data connection handle
}
