% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapporteketDbConnection.R
\name{rapOpenDbConnection}
\alias{rapOpenDbConnection}
\title{Provide connection handle for data source at Rapporteket}
\usage{
rapOpenDbConnection(registryName, dbType = "mysql")
}
\arguments{
\item{registryName}{String id used for the registry in global configuration
file from which information on the database connection is provided}

\item{dbType}{String providing type of data source, one of
"mysql" and "mssql". Defaults to "mysql"}
}
\value{
con Data source connection object

drv DBIDriver object
}
\description{
Generic to registries, handle the data source connections, including
usernames and passwords needed to open these connections
}
