% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{readAutoReportData}
\alias{readAutoReportData}
\title{Read automated report metadata}
\usage{
readAutoReportData(fileName = "autoReport.yml",
  packageName = "rapbase")
}
\arguments{
\item{fileName}{String defining name of the yaml configuration file. Default
'autoReport.yml'}

\item{packageName}{String defining the package in which the above
configuration file resides. A configuration file within an R-package is
only used in case the environmental variable 'R_RAP_CONFIG_PATH' is not
defined (empty)}
}
\value{
a list of yaml data
}
\description{
Read automated report metadata
}
\examples{
readAutoReportData()
}
