% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{readLog}
\alias{readLog}
\title{Read log entries}
\usage{
readLog(type, name = "")
}
\arguments{
\item{type}{Character string defining which log to request data from. Must be
one of \code{c("app", "report")}.}

\item{name}{Character string with registry filter. Default value is an empty
string that will return all log entries. If not empty its value must
correspond to an existing registry (\emph{i.e.} R package) name.}
}
\value{
A data frame of log entries
}
\description{
Internal function that provide log entries
}
\keyword{internal}
