% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{sanitizeLog}
\alias{sanitizeLog}
\title{Sanitize log entries that have reached end of life}
\usage{
sanitizeLog()
}
\value{
NULL on success
}
\description{
Sanitize log entries that have reached end of life
}
\keyword{internal}
