% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{selectByType}
\alias{selectByType}
\title{Select data by type from config (list)}
\usage{
selectByType(config, type)
}
\arguments{
\item{config}{List of configuration for automated reports}

\item{type}{Character vector with desired types}
}
\value{
list with config for reports of given types
}
\description{
Pick reports of a given type
}
