% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{theme}
\alias{theme}
\title{Theme of the app. Default is "flatly"}
\usage{
theme(theme = "flatly")
}
\arguments{
\item{theme}{}
}
\value{
a shinytheme object
}
\description{
Theme of the app. Default is "flatly"
}
