% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{theme}
\alias{theme}
\title{Theme of the app.}
\usage{
theme(theme = "flatly")
}
\arguments{
\item{theme}{Name of the theme. See ?shinythemes::shinytheme
for available themes.}
}
\value{
a shinytheme object
}
\description{
This is a wrapper around shinythemes::shinytheme
}
