% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{theme}
\alias{theme}
\title{Theme of the app.}
\usage{
theme(theme = "flatly", version = 3)
}
\arguments{
\item{theme}{Name of the theme. See bslib::bootswatch_themes()
for available themes.}
}
\value{
a shinytheme object
}
\description{
This is a wrapper around bslib::bs_theme
}
