% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{title}
\alias{title}
\title{Title on the top left of the app, including the logo}
\usage{
title(regTitle = "rapbase")
}
\arguments{
\item{regTitle}{Title of the app}
}
\value{
a div containing the logo and the title
}
\description{
Title on the top left of the app, including the logo
}
