% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAttribute.R
\name{userAttribute}
\alias{userAttribute}
\alias{getUserEmail}
\alias{getUserFullName}
\alias{getUserGroups}
\alias{getUserName}
\alias{getUserPhone}
\alias{getUserReshId}
\alias{getUserRole}
\title{User attributes in container apps running behind shinyproxy}
\usage{
userAttribute(unit = NULL, map_orgname = NULL)

getUserEmail(...)

getUserFullName(...)

getUserGroups(...)

getUserName(...)

getUserPhone(...)

getUserReshId(...)

getUserRole(...)
}
\arguments{
\item{unit}{Integer providing the look-up unit id. Default value is NULL in
which case all privileges for \code{group} are returned.}

\item{map_orgname}{A data.frame containing two columns:
\describe{
 \item{UnitId}{unit ids}
 \item{orgname}{corresponding organization names}
 }}
}
\value{
Invisibly a list of user metadata and privileges:
  \describe{
    \item{name}{The username for whom the privileges apply.}
    \item{fullName}{User full name}
    \item{phone}{User phone number}
    \item{email}{User email}
    \item{unit}{Unit id under which the privileges are defined.}
    \item{org}{Organization id for the user.}
    \item{role}{Role of the user.}
    \item{orgName}{Name of the organization as defined under the unit id.}
  }

String with user attribute
}
\description{
For apps running as containers particular environment variables must be
defined for an orderly handling of dynamic user privileges. This function
makes use of environmental variables defined by shinyproxy to provide
available privileges for the shiny application.

These are helper function for \code{\link{userInfo}}. When used without a
shiny session object calls to these functions is made without any arguments.
If redefining contexts is needed, please use \code{\link{userInfo}} instead.
}
\examples{
\donttest{
try(getUserEmail())
}
}
