% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReportFuns.R
\name{writeAutoReportData}
\alias{writeAutoReportData}
\title{Write automated report metadata}
\usage{
writeAutoReportData(fileName = "autoReport.yml", config,
  packageName = "rapbase")
}
\arguments{
\item{fileName}{String defining name of the yaml configuration file. Default
'autoReport.yml'}

\item{config}{a list of yaml configuration}

\item{packageName}{String defining the package in which the above
configuration file resides. A configuration file within an R-package is
only used in case the environmental variable 'R_RAP_CONFIG_PATH' is not
defined (empty)}
}
\description{
Write automated report metadata
}
\examples{
\donttest{
# Example depend on environment variable R_RAP_CONFIG_PATH being set
config <- readAutoReportData()
writeAutoReportData(config = config)
}
}
